atualização exercicios

+++ ListaExercicios01.Exercicio01/Program.cs
@@ -8,17 +8,17 @@
 
             Console.Write("Digite a largura da sua caixa: ");
 
-            int largura = Convert.ToInt32(Console.ReadLine());
+            double largura = Convert.ToInt32(Console.ReadLine());
 
             Console.Write("Digite o comprimento da sua caixa: ");
 
-            int comprimento = Convert.ToInt32(Console.ReadLine());
+            double comprimento = Convert.ToInt32(Console.ReadLine());
 
             Console.Write("Digite a altura da sua caixa: ");
             
-            int altura = Convert.ToInt32(Console.ReadLine());
+            double altura = Convert.ToInt32(Console.ReadLine());
 
-            int volume = largura * comprimento * altura;
+            double volume = largura * comprimento * altura;
 
             Console.Write("O volume da sua caixa é: " + volume);
